admin statistics and orfanizer stats are added

+++ app/Http/Controllers/admin/EventController.php
@@ -17,7 +17,7 @@ class EventController extends Controller
      */
     public function index()
     {
-
+        $user = Auth::user();
         $categories = Category::all();
         $events = Event::orderBy("created_at", "desc")->paginate(10);
         return view("events.index", compact("events", "categories"));
@@ -55,6 +55,9 @@ class EventController extends Controller
      */
     public function show(Event $event)
     {
+        if (Auth::id() !== $event->user_id) {
+            abort(403, 'Permission denied');
+        }
         return view("events.show", compact("event"));
     }
 
@@ -63,6 +66,9 @@ class EventController extends Controller
      */
     public function edit(Event $event)
     {
+        if (Auth::id() !== $event->user_id) {
+            abort(403, 'Permission denied');
+        }
         $categories = Category::all();
         return view("events.edit", compact("event", "categories"));
     }
@@ -73,6 +79,9 @@ class EventController extends Controller
     public function update(Request $request, Event $event)
     {
         if (Gate::denies('edit events')) {
+            abort(403, 'Permission denied');
+        }
+        if (Auth::id() !== $event->user_id) {
             abort(403, 'Permission denied');
         }
         $validated = $request->validate([
@@ -101,6 +110,9 @@ class EventController extends Controller
         if (Gate::denies('delete events')) {
             abort(403, 'Permission denied');
         }
+        if (Auth::id() !== $event->user_id) {
+            abort(403, 'Permission denied');
+        }
         $event->delete();
         session()->flash('success', "Event is deleted successfully");
 
@@ -111,7 +123,9 @@ class EventController extends Controller
         if (Gate::denies('approve events')) {
             abort(403, 'Permission denied');
         }
-
+        if (Auth::id() !== $event->user_id) {
+            abort(403, 'Permission denied');
+        }
         $event->update(['status' => 'approved']);
 
         return redirect()->back()->with('success', "{$event->title} is approved");
@@ -122,7 +136,9 @@ class EventController extends Controller
         if (Gate::denies('deny events')) {
             abort(403, 'Permission denied');
         }
-
+        if (Auth::id() !== $event->user_id) {
+            abort(403, 'Permission denied');
+        }
         $event->update(['status' => 'denied']);
 
         return redirect()->back()->with('success', "{$event->title} is denied");
@@ -131,6 +147,9 @@ class EventController extends Controller
     public function approveReservation(Reservation $reservation)
     {
         if (Gate::denies('approve reservations')) {
+            abort(403, 'Permission denied');
+        }
+        if (Auth::id() !== $reservation->user_id) {
             abort(403, 'Permission denied');
         }
 
@@ -151,6 +170,9 @@ class EventController extends Controller
     public function denyReservation(Reservation $reservation)
     {
         if (Gate::denies('deny reservations')) {
+            abort(403, 'Permission denied');
+        }
+        if (Auth::id() !== $reservation->user_id) {
             abort(403, 'Permission denied');
         }
